working on cart

+++ db/migrate/20190415175612_create_items.rb
@@ -3,7 +3,7 @@ class CreateItems < ActiveRecord::Migration[5.2]
     create_table :items do |t|
       t.string :name
       t.text :description
-      t.integer :cost
+      t.decimal :cost
       t.text :image_url
 
       t.timestamps
